+++ cel/src/threadpool.c
@@ -298,10 +298,10 @@ static CelThreadPool *cel_threadpool_wait_pool(void)
     int waittime;
     BOOL have_relayed_thread_marker = FALSE;
 
-    long local_max_unused_threads = cel_atomic_load(&max_unused_threads);
-    long local_max_idle_time = cel_atomic_load(&max_idle_time);
+    long local_max_unused_threads = (long)cel_atomic_load(&max_unused_threads);
+    long local_max_idle_time = (long)cel_atomic_load(&max_idle_time);
     long local_wakeup_thread_serial;
-    long last_wakeup_thread_serial = cel_atomic_load(&wakeup_thread_serial);
+    long last_wakeup_thread_serial = (long)cel_atomic_load(&wakeup_thread_serial);
 
     cel_atomic_increment(&num_unused_threads, 1);
     do
@@ -333,7 +333,7 @@ static CelThreadPool *cel_threadpool_wait_pool(void)
         }
         if (thread_pool == wakeup_thread_marker)
         {
-            local_wakeup_thread_serial = cel_atomic_load(&wakeup_thread_serial);
+            local_wakeup_thread_serial = (long)cel_atomic_load(&wakeup_thread_serial);
             if (last_wakeup_thread_serial  == local_wakeup_thread_serial)
             {
                 if (!have_relayed_thread_marker)
@@ -346,10 +346,10 @@ static CelThreadPool *cel_threadpool_wait_pool(void)
             }
             else
             {
-                local_max_unused_threads = cel_atomic_load(&max_unused_threads);
-                local_max_idle_time = cel_atomic_load(&max_idle_time);
+                local_max_unused_threads = (long)cel_atomic_load(&max_unused_threads);
+                local_max_idle_time = (long)cel_atomic_load(&max_idle_time);
                 last_wakeup_thread_serial = 
-                    cel_atomic_load(&wakeup_thread_serial);
+                    (long)cel_atomic_load(&wakeup_thread_serial);
 
                 have_relayed_thread_marker = FALSE;
             }
@@ -379,7 +379,7 @@ void cel_threadpool_set_max_unused_threads(int max_threads)
 
     max_threads = (max_threads < -1 ? -1 : max_threads);
     cel_atomic_store(&max_unused_threads, (long)max_threads);
-    if ((i = num_unused_threads) > 0)
+    if ((i = (int)num_unused_threads) > 0)
         cel_atomic_increment(&wakeup_thread_serial, 1);
 
     cel_asyncqueue_lock(unused_thread_queue);
@@ -393,12 +393,12 @@ void cel_threadpool_set_max_unused_threads(int max_threads)
 
 int cel_threadpool_get_max_unused_threads(void)
 {
-    return cel_atomic_load(&max_unused_threads);
+    return (int)cel_atomic_load(&max_unused_threads);
 }
 
 int cel_threadpool_get_num_unused_threads(void)
 {
-    return cel_atomic_load(&num_unused_threads);
+    return (int)cel_atomic_load(&num_unused_threads);
 }
 
 void cel_threadpool_stop_unused_threads(void)
@@ -406,7 +406,7 @@ void cel_threadpool_stop_unused_threads(void)
     int i, oldval = cel_threadpool_get_max_unused_threads();
 
     cel_threadpool_set_max_unused_threads(0);
-    if ((i = num_unused_threads) > 0)
+    if ((i = (int)num_unused_threads) > 0)
         cel_atomic_increment(&wakeup_thread_serial, 1);
 
     cel_asyncqueue_lock(unused_thread_queue);
@@ -425,7 +425,7 @@ void cel_threadpool_set_max_idle_time(int interval)
     int i;
 
     cel_atomic_store(&max_idle_time, (long)interval);
-    if ((i = num_unused_threads) > 0)
+    if ((i = (int)num_unused_threads) > 0)
         cel_atomic_increment(&wakeup_thread_serial, 1);
 
     cel_asyncqueue_lock(unused_thread_queue);
@@ -439,5 +439,5 @@ void cel_threadpool_set_max_idle_time(int interval)
 
 int cel_threadpool_get_max_idle_time(void)
 {
-    return cel_atomic_load(&max_idle_time);
+    return (int)cel_atomic_load(&max_idle_time);
 }
